Feature/index scan limit multiple rules (#2966)
* Support limit for lookup.

* Add multiple rules for index scan.

+++ src/graph/context/ast/QueryAstContext.h
@@ -113,7 +113,7 @@ struct LookupContext final : public AstContext {
   bool dedup{false};
   bool isEmptyResultSet{false};
   int32_t schemaId{-1};
-  int32_t limit{-1};
+  int64_t limit{-1};
   Expression* filter{nullptr};
   YieldColumns* yieldExpr{nullptr};
   std::vector<std::string> idxReturnCols;
